JLG: CTMART-1765 More Error handling framework examples

+++ src/main/java/com/bcmc/xor/flare/client/api/web/rest/ServerResource.java
@@ -1,17 +1,18 @@
 package com.bcmc.xor.flare.client.api.web.rest;
 
+import com.bcmc.xor.flare.client.api.domain.auth.User;
 import com.bcmc.xor.flare.client.api.domain.server.TaxiiServer;
 import com.bcmc.xor.flare.client.api.domain.server.TemporaryServer;
 import com.bcmc.xor.flare.client.api.security.SecurityUtils;
 import com.bcmc.xor.flare.client.api.service.ServerService;
+import com.bcmc.xor.flare.client.api.service.UserService;
 import com.bcmc.xor.flare.client.api.service.dto.ServerCredentialDTO;
 import com.bcmc.xor.flare.client.api.service.dto.ServerDTO;
 import com.bcmc.xor.flare.client.api.service.dto.ServersDTO;
-import com.bcmc.xor.flare.client.error.BadRequestAlertException;
-import com.bcmc.xor.flare.client.error.ErrorConstants;
-import com.bcmc.xor.flare.client.error.InternalServerErrorException;
+import com.bcmc.xor.flare.client.error.*;
 import com.bcmc.xor.flare.client.util.HeaderUtil;
 import com.bcmc.xor.flare.client.util.ResponseUtil;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
@@ -21,6 +22,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.Optional;
 
 @SuppressWarnings("unused")
 @RestController
@@ -29,9 +31,11 @@ class ServerResource {
 
     private static final Logger log = LoggerFactory.getLogger(ServerResource.class);
     private final ServerService serverService;
+    private final UserService userService;
 
-    public ServerResource(ServerService serverService) {
+    public ServerResource(ServerService serverService, UserService userService) {
         this.serverService = serverService;
+        this.userService = userService;
     }
 
     @PostMapping("/servers")
@@ -39,10 +43,8 @@ class ServerResource {
 
         // Check basic auth credentials; add them to ServerCredentialsUtils map
         if (serverDTO.getRequiresBasicAuth()) {
-            if (serverDTO.getUsername() == null || serverDTO.getPassword() == null
-                || serverDTO.getUsername().isEmpty() || serverDTO.getPassword().isEmpty()) {
-                throw new BadRequestAlertException("Server requires basic auth but username and/or password not supplied",
-                    serverDTO.getLabel(), ErrorConstants.ERR_BAD_REQUEST);
+            if (StringUtils.isBlank(serverDTO.getUsername()) || StringUtils.isBlank(serverDTO.getPassword())) {
+                return new ResponseEntity("Requires basic authentication is true.  Missing username and/or password.", HttpStatus.BAD_REQUEST);
             } else {
                 serverService.addServerCredential(serverDTO.getLabel(), serverDTO.getUsername(), serverDTO.getPassword());
             }
@@ -50,13 +52,12 @@ class ServerResource {
 
         TaxiiServer server = serverService.updateServer(serverDTO);
         if (server instanceof TemporaryServer) {
-            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(String.format("'%s' failed creation. Ensure server details are correct.", serverDTO.getLabel()), serverDTO.getLabel())).body(null);
+            return new ResponseEntity(String.format("'%s' failed creation. Ensure server details are correct.", serverDTO.getLabel()), HttpStatus.BAD_REQUEST);
         }
         return ResponseEntity
             .created(new URI("/api/servers/" + server.getId()))
             .headers(HeaderUtil.createAlert( String.format("'%s' created or updated", server.getLabel()), server.getLabel()))
             .body(server);
-
     }
 
     @GetMapping("/servers")
@@ -70,9 +71,13 @@ class ServerResource {
     }
 
     @DeleteMapping("/servers/{label}")
-    public ResponseEntity<Void> deleteServer(@PathVariable String label) {
+    public ResponseEntity<String> deleteServer(@PathVariable String label) {
+        log.debug("REST Request to delete server for '{}'", label);
+        if (!serverService.findOneByLabel(label).isPresent()) {
+            return new ResponseEntity<>("Unable to find server '" + label + "' for deletion", HttpStatus.BAD_REQUEST);
+        }
         serverService.deleteServer(label);
-        return ResponseEntity.ok().headers(HeaderUtil.createAlert(String.format("Deleted server '%s'", label), label)).build();
+        return new ResponseEntity<>("Successfully deleted server '" + label + "'", HttpStatus.OK);
     }
 
     @PostMapping("/servers/{label}/refresh")
@@ -92,7 +97,12 @@ class ServerResource {
 
     @DeleteMapping("/servers/{label}/credentials")
     public ResponseEntity<ServerDTO> deleteServerCredential(@PathVariable String label) {
-        log.debug("REST Request to delete server credential for user '{}' and server '{}'", SecurityUtils.getCurrentUserLogin(), label);
+        Optional<String> currentUser = SecurityUtils.getCurrentUserLogin();
+        log.debug("REST Request to delete server credential for user '{}' and server '{}'", currentUser, label);
+        Optional<User> user = userService.getUserWithAuthoritiesByLogin(currentUser.get());
+        if (!user.isPresent() || user.get().getServerCredentials().isEmpty()) {
+            return new ResponseEntity("Server credentials for user: '" + currentUser.get() + "' does not exist", HttpStatus.BAD_REQUEST);
+        }
         serverService.removeServerCredential(label);
         serverService.refreshServer(label);
         return getServer(label);
